refactor(csharp): remove language server dependency
