+++ fennara-cpp/include/fennara/csharp/project.hpp
@@ -6,6 +6,5 @@
 namespace fennara::csharp_support {
 
 godot::Dictionary inspect_project();
-godot::String diagnostics_unavailable_message(const godot::Dictionary &status);
 
 } // namespace fennara::csharp_support
