+++ fennara-cpp/src/tools/runtime_session/runtime_session.cpp
@@ -1,7 +1,7 @@
 #include "fennara/tools/runtime_session.hpp"
 
 #include "fennara/control_auth.hpp"
-#include "fennara/lsp/csharp_build.hpp"
+#include "fennara/csharp/build.hpp"
 #include "fennara/runtime/runtime_scene_preflight.hpp"
 #include "fennara/tools/validate_scene.hpp"
