+++ local/templates/fennara-guidelines.md
@@ -348,7 +348,7 @@ Use `script_diagnostics` when:
 - confirming a suspected parse/type/reference issue
 - errors only appear when opening scenes in Godot, such as invalid calls emitted by attached scripts while a scene is constructed
 
-Targeted `script_diagnostics` calls support `.gd` and `.gdshader` files. Targeted C# diagnostics are temporarily unavailable. To validate C#, finish the related edit set and call `script_diagnostics` once with `scan_project: true`; do not pass individual `.cs` paths. After Godot's initial editor filesystem scan completes, the plugin immediately starts worker-thread C# preparation: one isolated incremental build plus `csharp-ls` warmup. The isolated build still runs for a valid selected C# project when `csharp-ls` is unavailable; only targeted LSP warmup is skipped. C# tools wait for preparation without blocking the editor, and plugin-owned C# builds are serialized because they share Godot's intermediate MSBuild output. If C# source changes during the background build, it finishes normally and the next explicit project scan forces one refresh. `scan_project: true` scans project GDScript and shaders, then runs one cancellable incremental `dotnet build` for the selected C# project or solution graph instead of making one C# LSP request per file. The C# build writes final assemblies to isolated per-project output under `.godot/fennara/diagnostic_build`, so it does not update `.godot/mono/temp/bin/Debug` or trigger Godot editor assembly reload.
+Targeted `script_diagnostics` calls support `.gd` and `.gdshader` files. Targeted C# diagnostics are not supported. To validate C#, finish the related edit set and call `script_diagnostics` once with `scan_project: true`; do not pass individual `.cs` paths. After Godot's initial editor filesystem scan completes, the plugin immediately starts worker-thread C# preparation with one isolated incremental build. C# tools wait for preparation without blocking the editor, and plugin-owned C# builds are serialized because they share Godot's intermediate MSBuild output. If C# source changes during the background build, it finishes normally and the next explicit project scan forces one refresh. `scan_project: true` scans project GDScript and shaders, then runs one cancellable incremental `dotnet build` for the selected C# project or solution. The C# build writes final assemblies to isolated per-project output under `.godot/fennara/diagnostic_build`, so it does not update `.godot/mono/temp/bin/Debug` or trigger Godot editor assembly reload.
 
 For targeted `.gd` files, `script_diagnostics` also loads and instantiates project `.tscn` scenes in memory on the Godot main thread with Godot logger capture enabled. Script-related scene-load errors are attached to the matching script file with `source: "scene_load"` and `scene_path`, so the result says which scene triggered the script error. This is an in-memory diagnostic pass, not opening the editor UI or running the game scene.
